--- src/app/Shopping/Glasses.tsx
@@ -1,13 +1,0 @@
-import { View } from "react-native";
-import CartGlassesItem from "../../components/Shopping/CartGlassesItem";
-
-const Glasses = () => {
-  return (
-    <View>
-      <CartGlassesItem />
-      <CartGlassesItem />
-    </View>
-  )
-}
-
-export default Glasses
